requirements.txt
fastapi~=0.115.11
uvicorn~=0.34.0
pydantic~=2.10.6
python-jose~=3.4.0
motor~=3.7.0
pymongo~=4.11.3